requirements.txt
akshare
requests
pandas
streamlit
google-generativeai
python-dotenv
